sync

+++ cmd/ipfs/util/ulimit_test.go
@@ -1,3 +1,5 @@
+// +build !windows
+
 package util
 
 import (
@@ -73,7 +75,7 @@ func TestManageFdLimitWithEnvSet(t *testing.T) {
 	}
 
 	setMaxFds()
-	if maxFds != value {
+	if maxFds != uint64(value) {
 		t.Errorf("The maxfds is not set from IPFS_FD_MAX")
 	}
